Implmented spiral order enumeration of a 2d matrix
This is done by rotating the matrix and stripping off the first row.

To rotate the matrix we unpack and repack using * zip to flip the rows
and columns, then convert this back to a list (in python2 zip returned
a list but in python3 this is an enumerable). finally we reverse the
order using [::-1]